requirements.txt
flask==3.0.3
pyserial==3.5
flask-cors==4.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.3
